+++ R/proxy_example.R
@@ -3,8 +3,7 @@
 #' @description Launch an example to demonstrate how to use proxy method from \code{billboarder} in Shiny app.
 #'
 #' @param chart Chart type for which to see an example, possible values are \code{gauge}, \code{pie}, \code{bar}, 
-#' \code{bar2}, \code{line}, \code{line2}, \code{density}, \code{histogram}, \code{lollipop}, \code{stacked_bar}
-#'  or \code{transform} (for changing type of chart).
+#' \code{bar2}, \code{line}, \code{line2}, \code{density}, \code{histogram}, \code{lollipop}, \code{stacked_bar}.
 #'
 #' @export
 #' 
@@ -41,15 +40,12 @@
 #' # Histogram with ggplot2 diamonds
 #' proxy_example("histogram")
 #' 
-#' # Update chart type
-#' proxy_example("transform")
-#' 
 #' }
 proxy_example <- function(chart = "gauge") {
   chart <- match.arg(
     arg = chart,
     choices = c("gauge", "pie", "bar", "bar2", "line", "line2",
-                "density", "histogram", "transform", "stacked_bar", "lollipop"), 
+                "density", "histogram", "stacked_bar", "lollipop"), 
     several.ok = FALSE
   )
   path <- file.path("proxy_examples", chart)